Refactor 'rename' function to return bool in Data/Rename.php
Changed the return type of the 'rename' function in Data/Rename.php from mixed to bool to improve code reliability and predictability. Also, unnecessary comment lines were removed to enhance code readability, and operational logic was slightly re-adjusted by incorporating validation URLs check.

+++ src/Trait/Data/Rename.php
@@ -281,10 +281,12 @@ Trait Rename {
                     }
                 }
             }
-            File::delete($url_data_from);
-            File::delete($url_expose_from);
-            File::delete($url_object_from);
-            File::delete($url_validate_from);
+            if($from !== $to){
+                File::delete($url_data_from);
+                File::delete($url_expose_from);
+                File::delete($url_object_from);
+                File::delete($url_validate_from);
+            }
             return true;
         }
         return false;
